apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
supabase==2.9.1
python-dotenv==1.0.1
google-auth==2.37.0
requests>=2.31.0